+++ src/debugger/valueprint.h
@@ -17,7 +17,10 @@
 namespace dncdbg
 {
 
-HRESULT PrintValue(ICorDebugValue *pInputValue, std::string &output, bool escape = true);
+class Evaluator;
+
+HRESULT PrintValue(ICorDebugThread *pThread, Evaluator *pEvaluator,
+                   ICorDebugValue *pInputValue, std::string &output, bool escape = true);
 HRESULT GetNullableValue(ICorDebugValue *pValue, ICorDebugValue **ppValueValue, ICorDebugValue **ppHasValueValue);
 HRESULT PrintStringValue(ICorDebugValue *pValue, std::string &output);